Allow setting opts.from manually
I'd like to be able to manually set the `from` key, but it keeps getting overwritten. I don't really think this is a major change or heavily impacts or breaks any use cases, but it helps me if I want a consistent `from` that's very different from the file's path.